crew: use ExecPty where feasible
For commands which are run last before exiting crew, use ExecPty to
replace the current process image with the command to invoke running
with a pseudoterminal, rather than forking and pumping output through
the parent process. This improves the format of output over BlockPty,
which does not yet handle certain things such as terminal size
correctly.